+++ packages/core/test/actor.test.ts
@@ -1127,7 +1127,8 @@ describe('actors', () => {
           status: 'active',
           output: undefined,
           error: undefined
-        })
+        }),
+        getPersistedState: (s) => s
       };
 
       const pingMachine = createMachine({