+++ packages/core/test/invoke.test.ts
@@ -61,7 +61,6 @@ describe('invoke', () => {
           actors: {
             src: 'child';
             id: 'someService';
-            events: EventFrom<typeof childMachine>;
             logic: typeof childMachine;
           };
         },
@@ -2269,7 +2268,8 @@ describe('invoke', () => {
           output: undefined,
           error: undefined,
           context: 0
-        })
+        }),
+        getPersistedState: (s) => s
       };
 
       const countMachine = createMachine({
@@ -2309,7 +2309,8 @@ describe('invoke', () => {
           status: 'active',
           output: undefined,
           error: undefined
-        })
+        }),
+        getPersistedState: (s) => s
       };
 
       const pingMachine = createMachine({
@@ -3043,8 +3044,7 @@ describe('invoke', () => {
               onDone: {
                 guard: ({ event }) => {
                   // invoke ID should not be 'someSrc'
-                  const expectedType =
-                    'xstate.done.actor.(machine).a:invocation[0]';
+                  const expectedType = 'xstate.done.actor.(machine).a[0]';
                   expect(event.type).toEqual(expectedType);
                   return event.type === expectedType;
                 },
@@ -3111,7 +3111,7 @@ describe('invoke', () => {
       );
 
       expect(
-        createActor(machine).getSnapshot().children['machine.a:invocation[0]']
+        createActor(machine).getSnapshot().children['machine.a[0]']
       ).toBeDefined();
     }
   );
@@ -3225,8 +3225,8 @@ describe('invoke', () => {
     // check within a macrotask so all promise-induced microtasks have a chance to resolve first
     setTimeout(() => {
       expect(actual).toEqual([
-        'xstate.done.actor.(machine).first.fetch:invocation[0]',
-        'xstate.done.actor.(machine).second.fetch:invocation[0]'
+        'xstate.done.actor.(machine).first.fetch[0]',
+        'xstate.done.actor.(machine).second.fetch[0]'
       ]);
       done();
     }, 100);