+++ packages/core/test/rehydration.test.ts
@@ -23,7 +23,7 @@ describe('rehydration', () => {
       expect(service.getSnapshot().hasTag('foo')).toBe(true);
     });
 
-    it('should call exit actions when machine gets stopped immediately', () => {
+    it('should not call exit actions when machine gets stopped immediately', () => {
       const actual: string[] = [];
       const machine = createMachine({
         exit: () => actual.push('root'),
@@ -39,12 +39,11 @@ describe('rehydration', () => {
       const persistedState = JSON.stringify(actorRef.getPersistedState());
       actorRef.stop();
 
-      actual.length = 0;
       createActor(machine, { state: JSON.parse(persistedState) })
         .start()
         .stop();
 
-      expect(actual).toEqual(['a', 'root']);
+      expect(actual).toEqual([]);
     });
 
     it('should get correct result back from `can` immediately', () => {
@@ -92,7 +91,7 @@ describe('rehydration', () => {
       expect(service.getSnapshot().hasTag('foo')).toBe(true);
     });
 
-    it('should call exit actions when machine gets stopped immediately', () => {
+    it('should not call exit actions when machine gets stopped immediately', () => {
       const actual: string[] = [];
       const machine = createMachine({
         exit: () => actual.push('root'),
@@ -113,7 +112,7 @@ describe('rehydration', () => {
         .start()
         .stop();
 
-      expect(actual).toEqual(['active', 'root']);
+      expect(actual).toEqual([]);
     });
   });